+++ ElvUI_LuckyoneUI/Core/Database.lua
@@ -4,6 +4,8 @@ local L1UI, E, L, V, P, G = unpack(select(2, ...))
 function L1UI:Layout_Dragonflight(layout)
 	-- AB conversion
 	E.db.convertPages = true
+	-- Protect movers error
+	E.db.movers = E.db.movers or {}
 
 	-- General
 	E.db.general.afkChat = false
@@ -968,20 +970,18 @@ function L1UI:Layout_Dragonflight(layout)
 	E.db.unitframe.units.raid1.disableTargetGlow = true
 	E.db.unitframe.units.raid1.healPrediction.absorbStyle = 'WRAPPED'
 	E.db.unitframe.units.raid1.healPrediction.enable = true
-	E.db.unitframe.units.raid1.health.text_format = '[healthcolor][luckyone:health:percent]'
-	E.db.unitframe.units.raid1.health.xOffset = 0
-	E.db.unitframe.units.raid1.health.yOffset = 1
+	E.db.unitframe.units.raid1.health.text_format = ''
 	E.db.unitframe.units.raid1.horizontalSpacing = 1
 	E.db.unitframe.units.raid1.name.attachTextTo = 'Frame'
-	E.db.unitframe.units.raid1.name.position = 'TOP'
+	E.db.unitframe.units.raid1.name.position = 'CENTER'
 	E.db.unitframe.units.raid1.name.text_format = '[classcolor][name:veryshort]'
-	E.db.unitframe.units.raid1.name.yOffset = -1
+	E.db.unitframe.units.raid1.name.yOffset = 0
 	E.db.unitframe.units.raid1.numGroups = E.Retail and 4 or 2
-	E.db.unitframe.units.raid1.phaseIndicator.anchorPoint = 'BOTTOMRIGHT'
+	E.db.unitframe.units.raid1.phaseIndicator.anchorPoint = 'TOP'
 	E.db.unitframe.units.raid1.phaseIndicator.scale = 0.5
-	E.db.unitframe.units.raid1.phaseIndicator.xOffset = 1
+	E.db.unitframe.units.raid1.phaseIndicator.xOffset = 0
 	E.db.unitframe.units.raid1.power.enable = false
-	E.db.unitframe.units.raid1.pvpclassificationindicator.size = 18
+	E.db.unitframe.units.raid1.pvpclassificationindicator.enable = false
 	E.db.unitframe.units.raid1.raidicon.attachTo = 'TOPRIGHT'
 	E.db.unitframe.units.raid1.raidicon.attachToObject = 'Health'
 	E.db.unitframe.units.raid1.raidicon.size = 12
@@ -997,16 +997,20 @@ function L1UI:Layout_Dragonflight(layout)
 	E.db.unitframe.units.raid1.rdebuffs.stack.xOffset = 1
 	E.db.unitframe.units.raid1.rdebuffs.stack.yOffset = 0
 	E.db.unitframe.units.raid1.rdebuffs.yOffset = 1
-	E.db.unitframe.units.raid1.readycheckIcon.attachTo = 'Frame'
-	E.db.unitframe.units.raid1.readycheckIcon.position = 'CENTER'
+	E.db.unitframe.units.raid1.readycheckIcon.attachToObject = 'Frame'
+	E.db.unitframe.units.raid1.readycheckIcon.position = 'TOP'
+	E.db.unitframe.units.raid1.readycheckIcon.size = 14
+	E.db.unitframe.units.raid1.readycheckIcon.xOffset = 0
 	E.db.unitframe.units.raid1.readycheckIcon.yOffset = 0
+	E.db.unitframe.units.raid1.resurrectIcon.attachTo = 'RIGHT'
 	E.db.unitframe.units.raid1.resurrectIcon.size = 18
 	E.db.unitframe.units.raid1.roleIcon.attachTo = 'Frame'
 	E.db.unitframe.units.raid1.roleIcon.damager = false
-	E.db.unitframe.units.raid1.roleIcon.position = 'BOTTOMLEFT'
+	E.db.unitframe.units.raid1.roleIcon.position = 'LEFT'
 	E.db.unitframe.units.raid1.roleIcon.size = 12
 	E.db.unitframe.units.raid1.roleIcon.xOffset = 0
 	E.db.unitframe.units.raid1.roleIcon.yOffset = 0
+	E.db.unitframe.units.raid1.summonIcon.attachTo = 'RIGHT'
 	E.db.unitframe.units.raid1.summonIcon.size = 18
 	E.db.unitframe.units.raid1.threatStyle = 'NONE'
 	E.db.unitframe.units.raid1.verticalSpacing = 1
@@ -1029,11 +1033,14 @@ function L1UI:Layout_Dragonflight(layout)
 	E.db.unitframe.units.raid3.healPrediction.absorbStyle = 'WRAPPED'
 	E.db.unitframe.units.raid3.health.text_format = ''
 	E.db.unitframe.units.raid3.horizontalSpacing = 1
+	E.db.unitframe.units.raid3.name.attachTextTo = 'Frame'
+	E.db.unitframe.units.raid3.name.position = 'CENTER'
 	E.db.unitframe.units.raid3.name.text_format = '[classcolor][name:veryshort]'
-	E.db.unitframe.units.raid3.phaseIndicator.anchorPoint = 'BOTTOMRIGHT'
+	E.db.unitframe.units.raid3.name.yOffset = 0
+	E.db.unitframe.units.raid3.phaseIndicator.anchorPoint = 'LEFT'
 	E.db.unitframe.units.raid3.phaseIndicator.scale = 0.5
-	E.db.unitframe.units.raid3.phaseIndicator.xOffset = 1
-	E.db.unitframe.units.raid3.pvpclassificationindicator.size = 18
+	E.db.unitframe.units.raid3.phaseIndicator.xOffset = 0
+	E.db.unitframe.units.raid3.pvpclassificationindicator.enable = false
 	E.db.unitframe.units.raid3.raidicon.attachTo = 'TOPRIGHT'
 	E.db.unitframe.units.raid3.raidicon.attachToObject = 'Health'
 	E.db.unitframe.units.raid3.raidicon.size = 12
@@ -1048,26 +1055,24 @@ function L1UI:Layout_Dragonflight(layout)
 	E.db.unitframe.units.raid3.rdebuffs.size = 14
 	E.db.unitframe.units.raid3.rdebuffs.stack.position = 'CENTER'
 	E.db.unitframe.units.raid3.rdebuffs.stack.yOffset = 0
-	E.db.unitframe.units.raid3.readycheckIcon.position = 'TOPRIGHT'
-	E.db.unitframe.units.raid3.readycheckIcon.xOffset = -1
+	E.db.unitframe.units.raid3.readycheckIcon.attachToObject = 'Frame'
+	E.db.unitframe.units.raid3.readycheckIcon.position = 'RIGHT'
+	E.db.unitframe.units.raid3.readycheckIcon.xOffset = 0
 	E.db.unitframe.units.raid3.readycheckIcon.yOffset = 0
-	E.db.unitframe.units.raid3.resurrectIcon.attachTo = 'LEFT'
+	E.db.unitframe.units.raid3.resurrectIcon.attachTo = 'RIGHT'
 	E.db.unitframe.units.raid3.resurrectIcon.size = 18
 	E.db.unitframe.units.raid3.roleIcon.attachTo = 'Frame'
 	E.db.unitframe.units.raid3.roleIcon.damager = false
 	E.db.unitframe.units.raid3.roleIcon.enable = true
-	E.db.unitframe.units.raid3.roleIcon.position = 'BOTTOMLEFT'
+	E.db.unitframe.units.raid3.roleIcon.position = 'LEFT'
 	E.db.unitframe.units.raid3.roleIcon.size = 12
 	E.db.unitframe.units.raid3.roleIcon.xOffset = 0
 	E.db.unitframe.units.raid3.roleIcon.yOffset = 0
-	E.db.unitframe.units.raid3.summonIcon.attachTo = 'LEFT'
+	E.db.unitframe.units.raid3.summonIcon.attachTo = 'RIGHT'
 	E.db.unitframe.units.raid3.summonIcon.size = 18
 	E.db.unitframe.units.raid3.threatStyle = 'NONE'
 	E.db.unitframe.units.raid3.verticalSpacing = 1
 	E.db.unitframe.units.raid3.visibility = E.Retail and '[@raid31,noexists] hide;show' or '[@raid26,noexists] hide;show'
-
-	-- Protect movers error
-	E.db.movers = E.db.movers or {}
 
 	-- Shared movers
 	E.db.movers.AlertFrameMover = 'TOP,ElvUIParent,TOP,0,-178'
@@ -1191,18 +1196,21 @@ function L1UI:Layout_Dragonflight(layout)
 		E.db.unitframe.units.raid1.buffIndicator.enable = false
 		E.db.unitframe.units.raid1.height = 38
 		E.db.unitframe.units.raid1.rdebuffs.xOffset = 33
+		E.db.unitframe.units.raid1.rdebuffs.yOffset = 12
 		E.db.unitframe.units.raid1.width = 82
 
 		-- Main Raid2
 		E.db.unitframe.units.raid2.buffIndicator.enable = false
 		E.db.unitframe.units.raid2.height = 38
 		E.db.unitframe.units.raid2.rdebuffs.xOffset = 33
+		E.db.unitframe.units.raid2.rdebuffs.yOffset = 12
 		E.db.unitframe.units.raid2.width = 82
 
 		-- Main Raid3
 		E.db.unitframe.units.raid3.buffIndicator.enable = false
 		E.db.unitframe.units.raid3.height = 24
 		E.db.unitframe.units.raid3.rdebuffs.xOffset = 33
+		E.db.unitframe.units.raid3.rdebuffs.yOffset = 5
 		E.db.unitframe.units.raid3.width = 82
 	elseif layout == 'healing' then
 		-- Healing growth directions
@@ -1270,18 +1278,21 @@ function L1UI:Layout_Dragonflight(layout)
 		E.db.unitframe.units.raid1.buffIndicator.enable = true
 		E.db.unitframe.units.raid1.height = 40
 		E.db.unitframe.units.raid1.rdebuffs.xOffset = 39
+		E.db.unitframe.units.raid1.rdebuffs.yOffset = 13
 		E.db.unitframe.units.raid1.width = 94
 
 		-- Healing Raid2
 		E.db.unitframe.units.raid2.buffIndicator.enable = true
 		E.db.unitframe.units.raid2.height = 40
 		E.db.unitframe.units.raid2.rdebuffs.xOffset = 39
+		E.db.unitframe.units.raid2.rdebuffs.yOffset = 13
 		E.db.unitframe.units.raid2.width = 94
 
 		-- Healing Raid3
 		E.db.unitframe.units.raid3.buffIndicator.enable = true
 		E.db.unitframe.units.raid3.height = 23
 		E.db.unitframe.units.raid3.rdebuffs.xOffset = 39
+		E.db.unitframe.units.raid3.rdebuffs.yOffset = 4
 		E.db.unitframe.units.raid3.width = 94
 	end
 end
@@ -1290,6 +1301,8 @@ end
 function L1UI:Layout_Shadowlands(layout)
 	-- AB conversion
 	E.db.convertPages = true
+	-- Protect movers error
+	E.db.movers = E.db.movers or {}
 
 	-- General
 	E.db.general.afkChat = false
@@ -2161,9 +2174,6 @@ function L1UI:Layout_Shadowlands(layout)
 	E.db.unitframe.units.raid1.growthDirection = 'RIGHT_DOWN'
 	E.db.unitframe.units.raid2.growthDirection = 'RIGHT_DOWN'
 	E.db.unitframe.units.raid3.growthDirection = 'RIGHT_DOWN'
-
-	-- Protect movers error
-	E.db.movers = E.db.movers or {}
 
 	-- Shared movers
 	E.db.movers.AlertFrameMover = 'TOP,ElvUIParent,TOP,0,-135'